requirements.txt
numpy
librosa
matplotlib
seaborn
scikit-learn
torch
torchaudio
chatterbox-tts